Changed code organization while maintaining core functionality. Clarified the energy model in theory. Changed marginal weight calculations.
The major change was to coordinate the methods
associated with maintaining the grid and marginals
into a single class to prevent them getting out of
sync. Some conceptual changes were made to the way
the game state is tracked and updated.

Perhaps the biggest change was to remove the list
of integer intersections associated with each ship in
favor of recomputing the weight once as a function
of the array of squares that each ship occupies.

This change was intended to accommodate a wider
selection of weight functions that could not be
computed as a function of the number of intersections
alone.

This change may be leading to a slow down, since
I am now responsible for performing operations on
all squares occupied by a ship instead of the
intersection of the ship with the altered squares.

Ship named tuple became a class. I added an attribute
that stores all of the associated squares that the
ship occupies to avoid recomputing it.

Moved the visuals into their own file.

Added new weighting methods with a better description
of how they work.